+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Next Release
 ------------
 
 * Your contribution here.
+* Add Mongoid::History.disable and Mongoid::History.enabled? methods for global tracking disablement - [@johnnyshields](https://github.com/johnnyshields)
 * Add `:changes_method` that optionally overrides which method to call to collect changes - [@joelnordel](https://github.com/joelnordell).
 * [API Change] The `:destroy` action now stores trackers in the format `original=value, modified=nil` (previously it was the reverse) - [@johnnyshields](https://github.com/johnnyshields)
 * Support for polymorphic embedded classes - [@tstepp](https://github.com/tstepp)